requirements.txt
groq==0.31.0
huggingface-hub==0.34.4
Pillow==9.5.0
tokenizers==0.22.0
torch==2.6.0
torchvision==0.21.0
transformers==4.56.0